refactor(spells): extract custom spell schools editor state and types
- Move reactive state/handlers into CustomSpellSchoolsEditorState.svelte.ts
  (State suffix avoids the .svelte/.svelte.ts resolution collision when the
  component is imported from a .ts file)
- Move Props and row interface into CustomSpellSchoolsEditor.types.ts
- Thin the component to imports, props, state creation, and template
- Simplify duplicate-key detection to a Set
